拆分放置点识别
