+++ agent/go-service/aerosalvage/register.go
@@ -5,4 +5,5 @@ import maa "github.com/MaaXYZ/maa-framework-go/v4"
 // Register registers the Aerial Salvage custom recognition.
 func Register() {
 	maa.AgentServerRegisterCustomRecognition("AeroSalvageRecognition", &Recognition{})
+	maa.AgentServerRegisterCustomRecognition("AeroSalvagePlacementSiteRecognition", &PlacementSiteRecognition{})
 }
