+++ assets/resource/pipeline/AeroSalvage/Recognition.json
@@ -25,32 +25,6 @@
         "post_delay": 0,
         "rate_limit": 0
     },
-    "AeroSalvagePlacementSite": {
-        "desc": "气球放置位置",
-        "recognition": "ColorMatch",
-        "lower": [
-            235,
-            235,
-            125
-        ],
-        "upper": [
-            255,
-            255,
-            145
-        ]
-    },
-    "AeroSalvagePlacementSiteFeature": {
-        "desc": "气球放置位置 (FeatureMatch)",
-        "recognition": "FeatureMatch",
-        "template": "AeroSalvage/PlacementSite1.png",
-        "detector": "AKAZE",
-        "roi": [
-            497,
-            177,
-            283,
-            316
-        ]
-    },
     "AeroSalvagePlacementSiteTemplate": {
         "desc": "气球放置位置 (TemplateMatch)",
         "recognition": "TemplateMatch",
